created new downloadtools module to house download tools
