+++ .gitignore
@@ -7,3 +7,4 @@ build/
 dist/
 *egg-info/
 *.swp
+.ropeproject
