+++ easybuild/framework/easyblock.py
@@ -1,5 +1,5 @@
-# #
-# Copyright 2009-2014 Ghent University
+##
+# Copyright 2009-2015 Ghent University
 #
 # This file is part of EasyBuild,
 # originally created by the HPC team of Ghent University (http://ugent.be/hpc/en),
@@ -21,7 +21,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with EasyBuild.  If not, see <http://www.gnu.org/licenses/>.
-# #
+##
 """
 Generic EasyBuild support for building and installing software.
 The EasyBlock class should serve as a base class for all easyblocks.
@@ -56,15 +56,16 @@ from easybuild.framework.easyconfig.easyconfig import fetch_parameter_from_easyc
 from easybuild.framework.easyconfig.easyconfig import get_easyblock_class, get_module_path, resolve_template
 from easybuild.framework.easyconfig.tools import get_paths_for
 from easybuild.framework.easyconfig.templates import TEMPLATE_NAMES_EASYBLOCK_RUN_STEP
+from easybuild.tools.downloadtools import PROTOCOL_MAP
 from easybuild.tools.build_details import get_build_stats
 from easybuild.tools.build_log import EasyBuildError, print_error, print_msg
 from easybuild.tools.config import build_option, build_path, get_log_filename, get_repository, get_repositorypath
 from easybuild.tools.config import install_path, log_path, read_only_installdir, source_paths
 from easybuild.tools.environment import restore_env
-from easybuild.tools.filetools import DEFAULT_CHECKSUM, PROTOCOL_MAP
-from easybuild.tools.filetools import adjust_permissions, apply_patch, convert_name, download_file, encode_class_name
+from easybuild.tools.filetools import adjust_permissions, apply_patch, convert_name, encode_class_name
+from easybuild.tools.downloadtools import download_file
 from easybuild.tools.filetools import extract_file, mkdir, read_file, rmtree2
-from easybuild.tools.filetools import write_file, compute_checksum, verify_checksum
+from easybuild.tools.filetools import write_file, compute_checksum, verify_checksum, DEFAULT_CHECKSUM
 from easybuild.tools.run import run_cmd
 from easybuild.tools.jenkins import write_to_xml
 from easybuild.tools.module_generator import ModuleGenerator
